+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/nullish-coalescing/fixture.ts
@@ -1,0 +1,5 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+function processNullishCoalesce(s?: string) {
+  let len = s ?? '';
+}