+++ packages/ast-spec/src/legacy-fixtures/basics/fixtures/type-import-type/fixture.ts
@@ -1,0 +1,4 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+type A = typeof import('A');
+type B = import('B').X<Y>;