+++ packages/ast-spec/src/legacy-fixtures/errorRecovery/fixtures/_error_/class-multiple-implements/fixture.ts
@@ -1,0 +1,3 @@
+// TODO: This fixture might be too large, and if so should be split up.
+
+class a implements b implements c {}