+++ packages/shared-fixtures/fixtures/javascript/arrowFunctions/error-not-arrow.src.js
@@ -1,1 +1,0 @@
-left = (aSize.width/2) - ()