+++ packages/shared-fixtures/fixtures/javascript/basics/do-while-statements.src.js
@@ -1,6 +1,0 @@
-do; while(1);
-
-var i = 0;
-do {
-   i += 1;
-} while (i < 5);