+++ packages/shared-fixtures/fixtures/javascript/experimentalObjectRestSpread/shorthand-properties.src.js
@@ -1,9 +1,0 @@
-var foo,
-    get,
-    set;
-
-var x = {
-    foo,
-    get,
-    ...set
-};