+++ packages/shared-fixtures/fixtures/tsx/generic-jsx-member-expression-private.src.tsx
@@ -1,1 +1,0 @@
-const foo = <this.#foo<string>/>