+++ packages/shared-fixtures/fixtures/typescript/basics/angle-bracket-type-assertion-arrow-function.src.ts
@@ -1,1 +1,0 @@
-var asserted2 = <any>((n) => { return n; });