+++ packages/shared-fixtures/fixtures/typescript/basics/import-export-equal-declaration.src.ts
@@ -1,1 +1,0 @@
-export import foo = require('bar');