+++ packages/shared-fixtures/fixtures/typescript/basics/type-parameters-comments.src.ts
@@ -1,3 +1,0 @@
-foo< /* comment 1 */ A /* comment 2 */ >();
-function bar< /* aaa */ A /* bbb */ >() { }
-function baz< /* aaa */ A /* bbb */ = Foo >() { }