+++ packages/shared-fixtures/fixtures/typescript/decorators/accessor-decorators/accessor-decorator-factory-static-member.src.ts
@@ -1,4 +1,0 @@
-class Other {
-    @foo({ baz: true })
-    static get bar() { return this._bar; }
-}